requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
azure-monitor-opentelemetry==1.6.4
